requirements.txt
streamlit
torch
numpy
pandas
opencv-python
Pillow
scikit-learn
matplotlib
tqdm
ffmpeg-python
gdown